Move copy link button to its own component; remove use of global and !important
