+++ src/lib/flows/share/steps/share-url.svelte
@@ -3,19 +3,16 @@
   import Button from '$lib/components/button/button.svelte';
   import DownloadIcon from '$lib/components/icons/Download.svelte';
   import downloadUrl from '$lib/utils/download-url';
-  import LinkIcon from '$lib/components/icons/Link.svelte';
   import XIcon from '$lib/components/icons/X.svelte';
   import FacebookIcon from '$lib/components/icons/Facebook.svelte';
   import ThreadsIcon from '$lib/components/icons/Threads.svelte';
   import TelegramIcon from '$lib/components/icons/Telegram.svelte';
   import WhatsAppIcon from '$lib/components/icons/WhatsApp.svelte';
   import EllipsisIcon from '$lib/components/icons/Ellipsis.svelte';
-  import CheckCircleIcon from '$lib/components/icons/CheckCircle.svelte';
-  import CopyIcon from '$lib/components/icons/Copy.svelte';
   import EmailIcon from '$lib/components/icons/Email.svelte';
-  import { fade } from 'svelte/transition';
   import { browser } from '$app/environment';
   import { type ShareOption } from '../share-steps';
+  import CopyLinkButton from '$lib/components/copy-link-button/copy-link-button.svelte';
 
   export let url: string = '';
   export let downloadableImageUrl: string | undefined = undefined;
@@ -25,15 +22,6 @@
   $: shareText = text || '';
 
   const shareSupported = browser && navigator.share;
-
-  let hovering = false;
-  let copySuccess = false;
-
-  function copyShareLink() {
-    navigator.clipboard.writeText(url);
-    copySuccess = true;
-    setTimeout(() => (copySuccess = false), 1000);
-  }
 
   const shareOptions: ShareOption[] = [
     {
@@ -70,14 +58,14 @@
       name: 'More',
       icon: EllipsisIcon,
       onClick() {
-        if (shareSupported) {
-          navigator.share({
-            text,
-            url,
-          });
-        } else {
-          copyShareLink();
+        if (!shareSupported) {
+          return;
         }
+
+        navigator.share({
+          text,
+          url,
+        });
       },
     },
   ];
@@ -102,39 +90,16 @@
         <div class="downloadable-image__card">
           <img src={downloadableImageUrl} alt="downloadable header" />
         </div>
-        <Button on:click={downloadImage} variant="normal" icon={DownloadIcon}>Download</Button>
+        <span class="downloadable-image__button">
+          <Button on:click={downloadImage} variant="normal" icon={DownloadIcon}>Download</Button>
+        </span>
       </div>
     {/if}
     <div class="share-options pixelated">
       <h2 class="pixelated">Share</h2>
       <p>{shareModalText}</p>
       <div class="share-options__options">
-        <Button
-          on:mouseenter={() => (hovering = true)}
-          on:focus={() => (hovering = true)}
-          on:mouseleave={() => (hovering = false)}
-          on:blur={() => (hovering = false)}
-          on:click={copyShareLink}
-        >
-          <div class="button-inner">
-            <div class="icon">
-              {#if copySuccess}
-                <span transition:fade={{ duration: 200 }}>
-                  <CheckCircleIcon style="fill: var(--color-positive)" />
-                </span>
-              {:else if hovering}
-                <span transition:fade={{ duration: 200 }}>
-                  <CopyIcon style="fill: currentColor" />
-                </span>
-              {:else}
-                <span transition:fade={{ duration: 200 }}
-                  ><LinkIcon style="fill: currentColor" /></span
-                >
-              {/if}
-            </div>
-            Copy link
-          </div>
-        </Button>
+        <CopyLinkButton {url} />
 
         {#each shareOptions as option}
           <Button
@@ -142,7 +107,8 @@
             href={option.href}
             on:click={option?.onClick?.bind(option)}
             variant="normal"
-            icon={option.icon}>{option.name}</Button
+            icon={option.icon}
+            justify="left">{option.name}</Button
           >
         {/each}
       </div>
@@ -168,7 +134,7 @@
     aspect-ratio: 16/9;
   }
 
-  .downloadable-image :global(button) {
+  .downloadable-image .downloadable-image__button {
     position: absolute;
     right: 1rem;
     bottom: -1rem;
@@ -179,41 +145,9 @@
   }
 
   .share-options__options {
-    display: flex;
-    flex-wrap: wrap;
+    display: grid;
+    grid-template-columns: auto auto;
     gap: 1rem;
     margin-top: 2rem;
-  }
-
-  .share-options__options :global(button),
-  .share-options__options :global(a) {
-    width: calc(50% - 1rem);
-  }
-
-  .share-options__options :global(button .inner),
-  .share-options__options :global(a .inner) {
-    /* yoiks */
-    justify-content: start !important;
-  }
-
-  .button-inner {
-    display: flex;
-    align-items: center;
-    gap: 0.125rem;
-    margin-left: -0.25rem;
-  }
-
-  .button-inner .icon {
-    height: 2rem;
-    width: 2rem;
-    display: inline-flex;
-    justify-content: center;
-    align-items: center;
-    position: relative;
-    border-radius: 1rem;
-  }
-
-  .button-inner .icon > * {
-    position: absolute;
   }
 </style>
